implementing design from sketch part 2

+++ app/src/main/java/com/techme/direction/Note.java
@@ -18,7 +18,7 @@ public class Note {
         return note_id;
     }
 
-    public void setNote_id(int note_id) {
+    public void setNote_id(long note_id) {
         this.note_id = note_id;
     }
